Add support for `dataflow` url parameter - no need to assign a value (simply add `&dataflow` - DataFlow shows separate DataFlow header - DataFlow header has three buttons each corresponding to a separate "page" - "Control Panels" page shows result of Chris' iot work - "Data Flow Creator" and "Data Stories" pages just show string labels - DataFlow pages don't show side- and bottom-navigation widgets - DataFlow-specific components in separate top-level `src/dataflow` folder - CLUE-specific components in separate top-level `src/clue` folder - Demo mode supported for DataFlow and CLUE - Teacher mode for DataFlow shows DataFlow header but existing teacher content
