+++ src/clue/components/clue-header.sass
@@ -1,0 +1,91 @@
+@import ../../components/vars
+
+$member-size: 18px
+
+.clue-header
+  position: fixed
+  top: 0
+  left: 0
+  right: 0
+  height: $header-height
+  background-color: $color1
+  display: flex
+  flex-direction: row
+  flex-wrap: nowrap
+  justify-content: flex-start
+  font-size: 14px
+
+  .info
+    flex-grow: 1
+    padding: 0 $padding
+    display: flex
+    align-items: center
+
+    .problem
+      font-weight: bold
+    .class
+
+  .version
+    padding-right: $padding
+    font-size: 0.9em
+
+  .group
+    flex-grow: 0
+    padding: 0 $padding
+    text-align: center
+    display: flex
+    flex-direction: row
+    align-items: center
+
+    .name
+      font-weight: bold
+      cursor: pointer
+      background-color: $color1-3
+      padding: $half-padding $padding
+
+    .members
+      display: flex
+      flex-direction: column
+      .row
+        display: flex
+        flex-direction: row
+        .member
+          font-size: 10px
+          line-height: $member-size
+          margin: 1px
+          padding: 3px
+          width: $member-size
+          height: $member-size
+          background-color: #2da343
+          color: white
+          &.nw
+            background-color: blue
+          &.ne
+            background-color: orange
+            border-top-right-radius: $border-radius
+          &.se
+            background-color: green
+            border-bottom-right-radius: $border-radius
+          &.sw
+            background-color: purple
+            border-bottom-left-radius: $border-radius
+          &.disconnected
+            background-color: #777
+          &.teacher
+            background-color: #EA6D2F !important
+          &.empty
+            background-color: $color1-4
+
+          .initials
+            margin-top: -3px
+
+  .user
+    flex-grow: 0
+    display: flex
+    align-items: center
+    margin-right: $double-margin
+
+    .name
+      padding: $half-padding $padding
+      background-color: $color1-3
+      font-weight: bold
