+++ src/components/demo/demo-creator.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 import { BaseComponent, IBaseProps } from "../base";
 import { InvestigationModelType } from "../../models/curriculum/investigation";
 import { ProblemModelType } from "../../models/curriculum/problem";
+import { urlParams } from "../../utilities/url-params";
 
 import "./demo-creator.sass";
 
@@ -88,10 +89,11 @@ export class DemoCreatorComponment extends BaseComponent<IProps, {}> {
 
   private createLink(userType: string, userIndex: number) {
     const { demo, unit } = this.stores;
+    const appStr = urlParams.dataflow !== undefined ? "&dataflow" : "";
     const fakeUser = `${userType}:${userIndex}`;
     const unitStr = unit.code ? `&unit=${unit.code}` : "";
     // tslint:disable-next-line:max-line-length
-    const href = `?appMode=demo&fakeClass=${demo.class.id}&fakeUser=${fakeUser}${unitStr}&problem=${demo.problemOrdinal}`;
+    const href = `?appMode=demo${appStr}&fakeClass=${demo.class.id}&fakeUser=${fakeUser}${unitStr}&problem=${demo.problemOrdinal}`;
     return (
       <li key={userIndex}>
         <a href={href} target="_blank">{userType} {userIndex}</a>
