+++ src/dataflow/components/dataflow-header.sass
@@ -1,0 +1,53 @@
+@import ../../components/vars
+
+$member-size: 18px
+
+.dataflow-header
+  position: fixed
+  top: 0
+  left: 0
+  right: 0
+  height: $header-height
+  background-color: $color1
+  display: flex
+  flex-wrap: nowrap
+  justify-content: space-between
+  font-size: 14px
+
+  .left
+    flex-grow: 0
+    padding: 0 $padding
+    display: flex
+    align-items: center
+    font-weight: bold
+    font-size: 16px
+
+  .middle
+    display: flex
+    align-items: center
+
+    button
+      // note: need focus/accessibility solution
+      outline: none
+
+  .right
+    flex-grow: 0
+    padding: 0 $padding
+    text-align: center
+    display: flex
+    align-items: center
+
+    .version
+      padding-right: $padding
+      font-size: 0.9em
+
+    .user
+      flex-grow: 0
+      display: flex
+      align-items: center
+      margin-right: $double-margin
+
+      .name
+        padding: $half-padding $padding
+        background-color: $color1-3
+        font-weight: bold
