+++ src/dataflow/components/dataflow-types.ts
@@ -1,0 +1,1 @@
+export type DataflowPanelType = "control-panels" | "flow-creator" | "data-stories";
